refactor: move about-panel function
