+++ src/main/about-panel.ts
